Moved memoryharness to enginetest package and exported various methods
