+++ enginetest/memory_engine_test.go
@@ -16,6 +16,8 @@ package enginetest_test
 
 import (
 	"fmt"
+	"github.com/dolthub/go-mysql-server/memory"
+	"github.com/dolthub/go-mysql-server/sql/expression"
 	"testing"
 
 	"github.com/dolthub/go-mysql-server/enginetest"
@@ -26,6 +28,14 @@ import (
 // Any engine test that relies on the correct implementation of the in-memory database belongs here. All test logic and
 // queries are declared in the exported enginetest package to make them usable by integrators, to validate the engine
 // against their own implementation.
+
+type indexBehaviorTestParams struct {
+	name              string
+	driverInitializer enginetest.IndexDriverInitalizer
+	nativeIndexes     bool
+}
+
+const testNumPartitions = 5
 
 var numPartitionsVals = []int{
 	1,
@@ -56,7 +66,7 @@ func TestQueries(t *testing.T) {
 					continue
 				}
 				testName := fmt.Sprintf("partitions=%d,indexes=%v,parallelism=%v", numPartitions, indexBehavior.name, parallelism)
-				harness := newMemoryHarness(testName, parallelism, numPartitions, indexBehavior.nativeIndexes, indexBehavior.driverInitializer)
+				harness := enginetest.NewMemoryHarness(testName, parallelism, numPartitions, indexBehavior.nativeIndexes, indexBehavior.driverInitializer)
 
 				t.Run(testName, func(t *testing.T) {
 					enginetest.TestQueries(t, harness)
@@ -68,7 +78,7 @@ func TestQueries(t *testing.T) {
 
 // TestQueriesSimple runs the canonical test queries against a single threaded index enabled harness.
 func TestQueriesSimple(t *testing.T) {
-	enginetest.TestQueries(t, newMemoryHarness("simple", 1, testNumPartitions, true, nil))
+	enginetest.TestQueries(t, enginetest.NewMemoryHarness("simple", 1, testNumPartitions, true, nil))
 }
 
 // Convenience test for debugging a single query. Unskip and set to the desired query.
@@ -85,7 +95,7 @@ func TestSingleQuery(t *testing.T) {
 
 	fmt.Sprintf("%v", test)
 
-	harness := newMemoryHarness("", 1, testNumPartitions, true, mergableIndexDriver)
+	harness := enginetest.NewMemoryHarness("", 1, testNumPartitions, true, mergableIndexDriver)
 	engine := enginetest.NewEngine(t, harness)
 	engine.Analyzer.Debug = true
 	engine.Analyzer.Verbose = true
@@ -138,7 +148,7 @@ func TestSingleScript(t *testing.T) {
 
 	fmt.Sprintf("%v", test)
 
-	harness := newMemoryHarness("", 1, testNumPartitions, true, nil)
+	harness := enginetest.NewMemoryHarness("", 1, testNumPartitions, true, nil)
 	engine := enginetest.NewEngine(t, harness)
 	engine.Analyzer.Debug = true
 	engine.Analyzer.Verbose = true
@@ -147,7 +157,7 @@ func TestSingleScript(t *testing.T) {
 }
 
 func TestBrokenQueries(t *testing.T) {
-	enginetest.RunQueryTests(t, newSkippingMemoryHarness(), enginetest.BrokenQueries)
+	enginetest.RunQueryTests(t, enginetest.NewSkippingMemoryHarness(), enginetest.BrokenQueries)
 }
 
 func TestVersionedQueries(t *testing.T) {
@@ -155,7 +165,7 @@ func TestVersionedQueries(t *testing.T) {
 		for _, indexInit := range indexBehaviors {
 			for _, parallelism := range parallelVals {
 				testName := fmt.Sprintf("partitions=%d,indexes=%v,parallelism=%v", numPartitions, indexInit.name, parallelism)
-				harness := newMemoryHarness(testName, parallelism, numPartitions, indexInit.nativeIndexes, indexInit.driverInitializer)
+				harness := enginetest.NewMemoryHarness(testName, parallelism, numPartitions, indexInit.nativeIndexes, indexInit.driverInitializer)
 
 				t.Run(testName, func(t *testing.T) {
 					enginetest.TestVersionedQueries(t, harness)
@@ -176,144 +186,277 @@ func TestQueryPlans(t *testing.T) {
 
 	for _, indexInit := range indexBehaviors {
 		t.Run(indexInit.name, func(t *testing.T) {
-			harness := newMemoryHarness(indexInit.name, 1, 2, indexInit.nativeIndexes, indexInit.driverInitializer)
+			harness := enginetest.NewMemoryHarness(indexInit.name, 1, 2, indexInit.nativeIndexes, indexInit.driverInitializer)
 			enginetest.TestQueryPlans(t, harness)
 		})
 	}
 }
 
 func TestQueryErrors(t *testing.T) {
-	enginetest.TestQueryErrors(t, newDefaultMemoryHarness())
+	enginetest.TestQueryErrors(t, enginetest.NewDefaultMemoryHarness())
 }
 
 func TestInfoSchema(t *testing.T) {
-	enginetest.TestInfoSchema(t, newMemoryHarness("default", 1, testNumPartitions, true, mergableIndexDriver))
+	enginetest.TestInfoSchema(t, enginetest.NewMemoryHarness("default", 1, testNumPartitions, true, mergableIndexDriver))
 }
 
 func TestColumnAliases(t *testing.T) {
-	enginetest.TestColumnAliases(t, newDefaultMemoryHarness())
+	enginetest.TestColumnAliases(t, enginetest.NewDefaultMemoryHarness())
 }
 
 func TestOrderByGroupBy(t *testing.T) {
-	enginetest.TestOrderByGroupBy(t, newDefaultMemoryHarness())
+	enginetest.TestOrderByGroupBy(t, enginetest.NewDefaultMemoryHarness())
 }
 
 func TestAmbiguousColumnResolution(t *testing.T) {
-	enginetest.TestAmbiguousColumnResolution(t, newDefaultMemoryHarness())
+	enginetest.TestAmbiguousColumnResolution(t, enginetest.NewDefaultMemoryHarness())
 }
 
 func TestInsertInto(t *testing.T) {
-	enginetest.TestInsertInto(t, newDefaultMemoryHarness())
+	enginetest.TestInsertInto(t, enginetest.NewDefaultMemoryHarness())
 }
 
 func TestInsertIntoErrors(t *testing.T) {
-	enginetest.TestInsertIntoErrors(t, newDefaultMemoryHarness())
+	enginetest.TestInsertIntoErrors(t, enginetest.NewDefaultMemoryHarness())
 }
 
 func TestReplaceInto(t *testing.T) {
-	enginetest.TestReplaceInto(t, newDefaultMemoryHarness())
+	enginetest.TestReplaceInto(t, enginetest.NewDefaultMemoryHarness())
 }
 
 func TestReplaceIntoErrors(t *testing.T) {
-	enginetest.TestReplaceIntoErrors(t, newDefaultMemoryHarness())
+	enginetest.TestReplaceIntoErrors(t, enginetest.NewDefaultMemoryHarness())
 }
 
 func TestUpdate(t *testing.T) {
-	enginetest.TestUpdate(t, newMemoryHarness("default", 1, testNumPartitions, true, mergableIndexDriver))
+	enginetest.TestUpdate(t, enginetest.NewMemoryHarness("default", 1, testNumPartitions, true, mergableIndexDriver))
 }
 
 func TestUpdateErrors(t *testing.T) {
-	enginetest.TestUpdateErrors(t, newMemoryHarness("default", 1, testNumPartitions, true, mergableIndexDriver))
+	enginetest.TestUpdateErrors(t, enginetest.NewMemoryHarness("default", 1, testNumPartitions, true, mergableIndexDriver))
 }
 
 func TestDeleteFrom(t *testing.T) {
-	enginetest.TestDelete(t, newMemoryHarness("default", 1, testNumPartitions, true, mergableIndexDriver))
+	enginetest.TestDelete(t, enginetest.NewMemoryHarness("default", 1, testNumPartitions, true, mergableIndexDriver))
 }
 
 func TestDeleteFromErrors(t *testing.T) {
-	enginetest.TestDeleteErrors(t, newMemoryHarness("default", 1, testNumPartitions, true, mergableIndexDriver))
+	enginetest.TestDeleteErrors(t, enginetest.NewMemoryHarness("default", 1, testNumPartitions, true, mergableIndexDriver))
 }
 
 func TestScripts(t *testing.T) {
-	enginetest.TestScripts(t, newDefaultMemoryHarness())
+	enginetest.TestScripts(t, enginetest.NewDefaultMemoryHarness())
 }
 
 func TestTriggers(t *testing.T) {
-	enginetest.TestTriggers(t, newDefaultMemoryHarness())
+	enginetest.TestTriggers(t, enginetest.NewDefaultMemoryHarness())
 }
 
 func TestTriggersErrors(t *testing.T) {
-	enginetest.TestTriggerErrors(t, newDefaultMemoryHarness())
+	enginetest.TestTriggerErrors(t, enginetest.NewDefaultMemoryHarness())
 }
 
 func TestCreateTable(t *testing.T) {
-	enginetest.TestCreateTable(t, newDefaultMemoryHarness())
+	enginetest.TestCreateTable(t, enginetest.NewDefaultMemoryHarness())
 }
 
 func TestDropTable(t *testing.T) {
-	enginetest.TestDropTable(t, newDefaultMemoryHarness())
+	enginetest.TestDropTable(t, enginetest.NewDefaultMemoryHarness())
 }
 
 func TestRenameTable(t *testing.T) {
-	enginetest.TestRenameTable(t, newDefaultMemoryHarness())
+	enginetest.TestRenameTable(t, enginetest.NewDefaultMemoryHarness())
 }
 
 func TestRenameColumn(t *testing.T) {
-	enginetest.TestRenameColumn(t, newDefaultMemoryHarness())
+	enginetest.TestRenameColumn(t, enginetest.NewDefaultMemoryHarness())
 }
 
 func TestAddColumn(t *testing.T) {
-	enginetest.TestAddColumn(t, newDefaultMemoryHarness())
+	enginetest.TestAddColumn(t, enginetest.NewDefaultMemoryHarness())
 }
 
 func TestModifyColumn(t *testing.T) {
-	enginetest.TestModifyColumn(t, newDefaultMemoryHarness())
+	enginetest.TestModifyColumn(t, enginetest.NewDefaultMemoryHarness())
 }
 
 func TestDropColumn(t *testing.T) {
-	enginetest.TestDropColumn(t, newDefaultMemoryHarness())
+	enginetest.TestDropColumn(t, enginetest.NewDefaultMemoryHarness())
 }
 
 func TestCreateForeignKeys(t *testing.T) {
-	enginetest.TestCreateForeignKeys(t, newDefaultMemoryHarness())
+	enginetest.TestCreateForeignKeys(t, enginetest.NewDefaultMemoryHarness())
 }
 
 func TestDropForeignKeys(t *testing.T) {
-	enginetest.TestDropForeignKeys(t, newDefaultMemoryHarness())
+	enginetest.TestDropForeignKeys(t, enginetest.NewDefaultMemoryHarness())
 }
 
 func TestExplode(t *testing.T) {
-	enginetest.TestExplode(t, newDefaultMemoryHarness())
+	enginetest.TestExplode(t, enginetest.NewDefaultMemoryHarness())
 }
 
 func TestReadOnly(t *testing.T) {
-	enginetest.TestReadOnly(t, newDefaultMemoryHarness())
+	enginetest.TestReadOnly(t, enginetest.NewDefaultMemoryHarness())
 }
 
 func TestViews(t *testing.T) {
-	enginetest.TestViews(t, newDefaultMemoryHarness())
+	enginetest.TestViews(t, enginetest.NewDefaultMemoryHarness())
 }
 
 func TestVersionedViews(t *testing.T) {
-	enginetest.TestVersionedViews(t, newDefaultMemoryHarness())
+	enginetest.TestVersionedViews(t, enginetest.NewDefaultMemoryHarness())
 }
 
 func TestNaturalJoin(t *testing.T) {
-	enginetest.TestNaturalJoin(t, newDefaultMemoryHarness())
+	enginetest.TestNaturalJoin(t, enginetest.NewDefaultMemoryHarness())
 }
 
 func TestNaturalJoinEqual(t *testing.T) {
-	enginetest.TestNaturalJoinEqual(t, newDefaultMemoryHarness())
+	enginetest.TestNaturalJoinEqual(t, enginetest.NewDefaultMemoryHarness())
 }
 
 func TestNaturalJoinDisjoint(t *testing.T) {
-	enginetest.TestNaturalJoinDisjoint(t, newDefaultMemoryHarness())
+	enginetest.TestNaturalJoinDisjoint(t, enginetest.NewDefaultMemoryHarness())
 }
 
 func TestInnerNestedInNaturalJoins(t *testing.T) {
-	enginetest.TestInnerNestedInNaturalJoins(t, newDefaultMemoryHarness())
+	enginetest.TestInnerNestedInNaturalJoins(t, enginetest.NewDefaultMemoryHarness())
 }
 
 func TestColumnDefaults(t *testing.T) {
-	enginetest.TestColumnDefaults(t, newDefaultMemoryHarness())
+	enginetest.TestColumnDefaults(t, enginetest.NewDefaultMemoryHarness())
+}
+
+func unmergableIndexDriver(dbs []sql.Database) sql.IndexDriver {
+	return memory.NewIndexDriver("mydb", map[string][]sql.DriverIndex{
+		"mytable": {
+			newUnmergableIndex(dbs, "mytable",
+				expression.NewGetFieldWithTable(0, sql.Int64, "mytable", "i", false)),
+			newUnmergableIndex(dbs, "mytable",
+				expression.NewGetFieldWithTable(1, sql.Text, "mytable", "s", false)),
+			newUnmergableIndex(dbs, "mytable",
+				expression.NewGetFieldWithTable(0, sql.Int64, "mytable", "i", false),
+				expression.NewGetFieldWithTable(1, sql.Text, "mytable", "s", false)),
+		},
+		"othertable": {
+			newUnmergableIndex(dbs, "othertable",
+				expression.NewGetFieldWithTable(0, sql.Text, "othertable", "s2", false)),
+			newUnmergableIndex(dbs, "othertable",
+				expression.NewGetFieldWithTable(1, sql.Text, "othertable", "i2", false)),
+			newUnmergableIndex(dbs, "othertable",
+				expression.NewGetFieldWithTable(0, sql.Text, "othertable", "s2", false),
+				expression.NewGetFieldWithTable(1, sql.Text, "othertable", "i2", false)),
+		},
+		"bigtable": {
+			newUnmergableIndex(dbs, "bigtable",
+				expression.NewGetFieldWithTable(0, sql.Text, "bigtable", "t", false)),
+		},
+		"floattable": {
+			newUnmergableIndex(dbs, "floattable",
+				expression.NewGetFieldWithTable(2, sql.Text, "floattable", "f64", false)),
+		},
+		"niltable": {
+			newUnmergableIndex(dbs, "niltable",
+				expression.NewGetFieldWithTable(0, sql.Int64, "niltable", "i", false)),
+		},
+		"one_pk": {
+			newUnmergableIndex(dbs, "one_pk",
+				expression.NewGetFieldWithTable(0, sql.Int8, "one_pk", "pk", false)),
+		},
+		"two_pk": {
+			newUnmergableIndex(dbs, "two_pk",
+				expression.NewGetFieldWithTable(0, sql.Int8, "two_pk", "pk1", false),
+				expression.NewGetFieldWithTable(1, sql.Int8, "two_pk", "pk2", false),
+			),
+		},
+	})
+}
+
+func mergableIndexDriver(dbs []sql.Database) sql.IndexDriver {
+	return memory.NewIndexDriver("mydb", map[string][]sql.DriverIndex{
+		"mytable": {
+			newMergableIndex(dbs, "mytable",
+				expression.NewGetFieldWithTable(0, sql.Int64, "mytable", "i", false)),
+			newMergableIndex(dbs, "mytable",
+				expression.NewGetFieldWithTable(1, sql.Text, "mytable", "s", false)),
+			newMergableIndex(dbs, "mytable",
+				expression.NewGetFieldWithTable(0, sql.Int64, "mytable", "i", false),
+				expression.NewGetFieldWithTable(1, sql.Text, "mytable", "s", false)),
+		},
+		"othertable": {
+			newMergableIndex(dbs, "othertable",
+				expression.NewGetFieldWithTable(0, sql.Text, "othertable", "s2", false)),
+			newMergableIndex(dbs, "othertable",
+				expression.NewGetFieldWithTable(1, sql.Text, "othertable", "i2", false)),
+			newMergableIndex(dbs, "othertable",
+				expression.NewGetFieldWithTable(0, sql.Text, "othertable", "s2", false),
+				expression.NewGetFieldWithTable(1, sql.Text, "othertable", "i2", false)),
+		},
+		"bigtable": {
+			newMergableIndex(dbs, "bigtable",
+				expression.NewGetFieldWithTable(0, sql.Text, "bigtable", "t", false)),
+		},
+		"floattable": {
+			newMergableIndex(dbs, "floattable",
+				expression.NewGetFieldWithTable(2, sql.Text, "floattable", "f64", false)),
+		},
+		"niltable": {
+			newMergableIndex(dbs, "niltable",
+				expression.NewGetFieldWithTable(0, sql.Int64, "niltable", "i", false)),
+		},
+		"one_pk": {
+			newMergableIndex(dbs, "one_pk",
+				expression.NewGetFieldWithTable(0, sql.Int8, "one_pk", "pk", false)),
+		},
+		"two_pk": {
+			newMergableIndex(dbs, "two_pk",
+				expression.NewGetFieldWithTable(0, sql.Int8, "two_pk", "pk1", false),
+				expression.NewGetFieldWithTable(1, sql.Int8, "two_pk", "pk2", false),
+			),
+		},
+	})
+}
+
+func newUnmergableIndex(dbs []sql.Database, tableName string, exprs ...sql.Expression) *memory.UnmergeableIndex {
+	db, table := findTable(dbs, tableName)
+	return &memory.UnmergeableIndex{
+		memory.MergeableIndex{
+			DB:         db.Name(),
+			DriverName: memory.IndexDriverId,
+			TableName:  tableName,
+			Tbl:        table.(*memory.Table),
+			Exprs:      exprs,
+		},
+	}
+}
+
+func newMergableIndex(dbs []sql.Database, tableName string, exprs ...sql.Expression) *memory.MergeableIndex {
+	db, table := findTable(dbs, tableName)
+	if db == nil {
+		return nil
+	}
+	return &memory.MergeableIndex{
+		DB:         db.Name(),
+		DriverName: memory.IndexDriverId,
+		TableName:  tableName,
+		Tbl:        table.(*memory.Table),
+		Exprs:      exprs,
+	}
+}
+
+func findTable(dbs []sql.Database, tableName string) (sql.Database, sql.Table) {
+	for _, db := range dbs {
+		names, err := db.GetTableNames(sql.NewEmptyContext())
+		if err != nil {
+			panic(err)
+		}
+		for _, name := range names {
+			if name == tableName {
+				table, _, _ := db.GetTableInsensitive(sql.NewEmptyContext(), name)
+				return db, table
+			}
+		}
+	}
+	return nil, nil
 }
