--- enginetest/memoryharness_test.go
@@ -1,268 +1,0 @@
-// Copyright 2020 Liquidata, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package enginetest_test
-
-import (
-	"context"
-
-	"github.com/dolthub/go-mysql-server/enginetest"
-	"github.com/dolthub/go-mysql-server/memory"
-	"github.com/dolthub/go-mysql-server/sql"
-	"github.com/dolthub/go-mysql-server/sql/expression"
-)
-
-type memoryHarness struct {
-	name                   string
-	parallelism            int
-	numTablePartitions     int
-	indexDriverInitializer indexDriverInitalizer
-	nativeIndexSupport     bool
-	session                sql.Session
-}
-
-type indexBehaviorTestParams struct {
-	name              string
-	driverInitializer indexDriverInitalizer
-	nativeIndexes     bool
-}
-
-const testNumPartitions = 5
-
-func newMemoryHarness(name string, parallelism int, numTablePartitions int, useNativeIndexes bool, indexDriverInitalizer indexDriverInitalizer) *memoryHarness {
-	return &memoryHarness{
-		name:                   name,
-		numTablePartitions:     numTablePartitions,
-		indexDriverInitializer: indexDriverInitalizer,
-		parallelism:            parallelism,
-		nativeIndexSupport:     useNativeIndexes,
-	}
-}
-
-func newDefaultMemoryHarness() *memoryHarness {
-	return newMemoryHarness("default", 1, testNumPartitions, false, nil)
-}
-
-func newSkippingMemoryHarness() *skippingMemoryHarness {
-	return &skippingMemoryHarness{
-		memoryHarness: *newDefaultMemoryHarness(),
-	}
-}
-
-var _ enginetest.Harness = (*memoryHarness)(nil)
-var _ enginetest.IndexDriverHarness = (*memoryHarness)(nil)
-var _ enginetest.IndexHarness = (*memoryHarness)(nil)
-var _ enginetest.VersionedDBHarness = (*memoryHarness)(nil)
-var _ enginetest.ForeignKeyHarness = (*memoryHarness)(nil)
-var _ enginetest.SkippingHarness = (*skippingMemoryHarness)(nil)
-
-type skippingMemoryHarness struct {
-	memoryHarness
-}
-
-func (s skippingMemoryHarness) SkipQueryTest(query string) bool {
-	return true
-}
-
-func (m *memoryHarness) SupportsNativeIndexCreation() bool {
-	return m.nativeIndexSupport
-}
-
-func (m *memoryHarness) SupportsForeignKeys() bool {
-	return true
-}
-
-func (m *memoryHarness) Parallelism() int {
-	return m.parallelism
-}
-
-func (m *memoryHarness) NewContext() *sql.Context {
-	if m.session == nil {
-		m.session = enginetest.NewBaseSession()
-	}
-
-	return sql.NewContext(
-		context.Background(),
-		sql.WithSession(m.session),
-	)
-}
-
-func (m *memoryHarness) NewTableAsOf(db sql.VersionedDatabase, name string, schema sql.Schema, asOf interface{}) sql.Table {
-	table := memory.NewPartitionedTable(name, schema, m.numTablePartitions)
-	if m.nativeIndexSupport {
-		table.EnablePrimaryKeyIndexes()
-	}
-	db.(*memory.HistoryDatabase).AddTableAsOf(name, table, asOf)
-	return table
-}
-
-func (m *memoryHarness) SnapshotTable(db sql.VersionedDatabase, name string, asOf interface{}) error {
-	// Nothing to do for this implementation: the NewTableAsOf method does all the work of creating the snapshot.
-	return nil
-}
-
-func (m *memoryHarness) IndexDriver(dbs []sql.Database) sql.IndexDriver {
-	if m.indexDriverInitializer != nil {
-		return m.indexDriverInitializer(dbs)
-	}
-	return nil
-}
-
-func (m *memoryHarness) NewDatabase(name string) sql.Database {
-	return memory.NewHistoryDatabase(name)
-}
-
-func (m *memoryHarness) NewTable(db sql.Database, name string, schema sql.Schema) (sql.Table, error) {
-	table := memory.NewPartitionedTable(name, schema, m.numTablePartitions)
-	if m.nativeIndexSupport {
-		table.EnablePrimaryKeyIndexes()
-	}
-	db.(*memory.HistoryDatabase).AddTable(name, table)
-	return table, nil
-}
-
-type indexDriverInitalizer func([]sql.Database) sql.IndexDriver
-
-func unmergableIndexDriver(dbs []sql.Database) sql.IndexDriver {
-	return memory.NewIndexDriver("mydb", map[string][]sql.DriverIndex{
-		"mytable": {
-			newUnmergableIndex(dbs, "mytable",
-				expression.NewGetFieldWithTable(0, sql.Int64, "mytable", "i", false)),
-			newUnmergableIndex(dbs, "mytable",
-				expression.NewGetFieldWithTable(1, sql.Text, "mytable", "s", false)),
-			newUnmergableIndex(dbs, "mytable",
-				expression.NewGetFieldWithTable(0, sql.Int64, "mytable", "i", false),
-				expression.NewGetFieldWithTable(1, sql.Text, "mytable", "s", false)),
-		},
-		"othertable": {
-			newUnmergableIndex(dbs, "othertable",
-				expression.NewGetFieldWithTable(0, sql.Text, "othertable", "s2", false)),
-			newUnmergableIndex(dbs, "othertable",
-				expression.NewGetFieldWithTable(1, sql.Text, "othertable", "i2", false)),
-			newUnmergableIndex(dbs, "othertable",
-				expression.NewGetFieldWithTable(0, sql.Text, "othertable", "s2", false),
-				expression.NewGetFieldWithTable(1, sql.Text, "othertable", "i2", false)),
-		},
-		"bigtable": {
-			newUnmergableIndex(dbs, "bigtable",
-				expression.NewGetFieldWithTable(0, sql.Text, "bigtable", "t", false)),
-		},
-		"floattable": {
-			newUnmergableIndex(dbs, "floattable",
-				expression.NewGetFieldWithTable(2, sql.Text, "floattable", "f64", false)),
-		},
-		"niltable": {
-			newUnmergableIndex(dbs, "niltable",
-				expression.NewGetFieldWithTable(0, sql.Int64, "niltable", "i", false)),
-		},
-		"one_pk": {
-			newUnmergableIndex(dbs, "one_pk",
-				expression.NewGetFieldWithTable(0, sql.Int8, "one_pk", "pk", false)),
-		},
-		"two_pk": {
-			newUnmergableIndex(dbs, "two_pk",
-				expression.NewGetFieldWithTable(0, sql.Int8, "two_pk", "pk1", false),
-				expression.NewGetFieldWithTable(1, sql.Int8, "two_pk", "pk2", false),
-			),
-		},
-	})
-}
-
-func mergableIndexDriver(dbs []sql.Database) sql.IndexDriver {
-	return memory.NewIndexDriver("mydb", map[string][]sql.DriverIndex{
-		"mytable": {
-			newMergableIndex(dbs, "mytable",
-				expression.NewGetFieldWithTable(0, sql.Int64, "mytable", "i", false)),
-			newMergableIndex(dbs, "mytable",
-				expression.NewGetFieldWithTable(1, sql.Text, "mytable", "s", false)),
-			newMergableIndex(dbs, "mytable",
-				expression.NewGetFieldWithTable(0, sql.Int64, "mytable", "i", false),
-				expression.NewGetFieldWithTable(1, sql.Text, "mytable", "s", false)),
-		},
-		"othertable": {
-			newMergableIndex(dbs, "othertable",
-				expression.NewGetFieldWithTable(0, sql.Text, "othertable", "s2", false)),
-			newMergableIndex(dbs, "othertable",
-				expression.NewGetFieldWithTable(1, sql.Text, "othertable", "i2", false)),
-			newMergableIndex(dbs, "othertable",
-				expression.NewGetFieldWithTable(0, sql.Text, "othertable", "s2", false),
-				expression.NewGetFieldWithTable(1, sql.Text, "othertable", "i2", false)),
-		},
-		"bigtable": {
-			newMergableIndex(dbs, "bigtable",
-				expression.NewGetFieldWithTable(0, sql.Text, "bigtable", "t", false)),
-		},
-		"floattable": {
-			newMergableIndex(dbs, "floattable",
-				expression.NewGetFieldWithTable(2, sql.Text, "floattable", "f64", false)),
-		},
-		"niltable": {
-			newMergableIndex(dbs, "niltable",
-				expression.NewGetFieldWithTable(0, sql.Int64, "niltable", "i", false)),
-		},
-		"one_pk": {
-			newMergableIndex(dbs, "one_pk",
-				expression.NewGetFieldWithTable(0, sql.Int8, "one_pk", "pk", false)),
-		},
-		"two_pk": {
-			newMergableIndex(dbs, "two_pk",
-				expression.NewGetFieldWithTable(0, sql.Int8, "two_pk", "pk1", false),
-				expression.NewGetFieldWithTable(1, sql.Int8, "two_pk", "pk2", false),
-			),
-		},
-	})
-}
-
-func newUnmergableIndex(dbs []sql.Database, tableName string, exprs ...sql.Expression) *memory.UnmergeableIndex {
-	db, table := findTable(dbs, tableName)
-	return &memory.UnmergeableIndex{
-		memory.MergeableIndex{
-			DB:         db.Name(),
-			DriverName: memory.IndexDriverId,
-			TableName:  tableName,
-			Tbl:        table.(*memory.Table),
-			Exprs:      exprs,
-		},
-	}
-}
-
-func newMergableIndex(dbs []sql.Database, tableName string, exprs ...sql.Expression) *memory.MergeableIndex {
-	db, table := findTable(dbs, tableName)
-	if db == nil {
-		return nil
-	}
-	return &memory.MergeableIndex{
-		DB:         db.Name(),
-		DriverName: memory.IndexDriverId,
-		TableName:  tableName,
-		Tbl:        table.(*memory.Table),
-		Exprs:      exprs,
-	}
-}
-
-func findTable(dbs []sql.Database, tableName string) (sql.Database, sql.Table) {
-	for _, db := range dbs {
-		names, err := db.GetTableNames(sql.NewEmptyContext())
-		if err != nil {
-			panic(err)
-		}
-		for _, name := range names {
-			if name == tableName {
-				table, _, _ := db.GetTableInsensitive(sql.NewEmptyContext(), name)
-				return db, table
-			}
-		}
-	}
-	return nil, nil
-}
